Move resolveFunction and evaluateScalarSubquery out of QueryAnalyzer.cpp
